fix(wave1): production build TDZ - break perf circular dep, reorder App.tsx refreshData, fix schema.sql index order
